requirements copy.txt
APScheduler==3.11.0
eventkit==1.0.3
ib-insync==0.9.86
loguru==0.7.3
nest-asyncio==1.6.0
numpy==2.3.3
pandas==2.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
six==1.17.0
tzdata==2025.2
tzlocal==5.3.1
